+++ alchemy/src/cloudflare/bundle/bundle-worker.ts
@@ -13,6 +13,7 @@ import {
 import { external, external_als } from "./external.ts";
 import { getNodeJSCompatMode } from "./nodejs-compat-mode.ts";
 import { nodeJsCompatPlugin } from "./nodejs-compat.ts";
+import { nodeJsImportWarningPlugin } from "./nodejs-import-warning-plugin.ts";
 import { wasmPlugin } from "./wasm-plugin.ts";
 
 export type NoBundleResult = {
@@ -25,10 +26,9 @@ export async function bundleWorkerScript<B extends Bindings>(
     entrypoint: string;
     compatibilityDate: string;
     compatibilityFlags: string[];
+    cwd: string;
   },
 ): Promise<string | NoBundleResult> {
-  const projectRoot = props.projectRoot ?? process.cwd();
-
   const nodeJsCompatMode = await getNodeJSCompatMode(
     props.compatibilityDate,
     props.compatibilityFlags,
@@ -99,7 +99,7 @@ export async function bundleWorkerScript<B extends Bindings>(
       ...(props.bundle || {}),
       conditions: ["workerd", "worker", "import", "module", "browser"],
       mainFields: ["module", "main"],
-      absWorkingDir: projectRoot,
+      absWorkingDir: props.cwd,
       keepNames: true, // Important for Durable Object classes
       loader: {
         ".sql": "text",
@@ -109,12 +109,14 @@ export async function bundleWorkerScript<B extends Bindings>(
       plugins: [
         wasmPlugin,
         ...(props.bundle?.plugins ?? []),
-        ...(nodeJsCompatMode === "v2" ? [await nodeJsCompatPlugin()] : []),
+        nodeJsCompatMode === "v2"
+          ? await nodeJsCompatPlugin()
+          : nodeJsImportWarningPlugin(nodeJsCompatMode),
         ...(props.bundle?.alias
           ? [
               createAliasPlugin({
                 alias: props.bundle?.alias,
-                projectRoot,
+                projectRoot: props.cwd,
               }),
             ]
           : []),