+++ alchemy/src/cloudflare/worker.ts
@@ -135,6 +135,12 @@ export interface BaseWorkerProps<
   /**
    * The root directory of the project
    */
+  cwd?: string;
+
+  /**
+   * The root directory of the project
+   * @deprecated Use `cwd` instead
+   */
   projectRoot?: string;
 
   /**
@@ -471,6 +477,12 @@ export type Worker<
      * The name of the worker
      */
     name: string;
+
+    /**
+     * The root directory of the project
+     * @default process.cwd()
+     */
+    cwd: string;
 
     /**
      * Time at which the worker was created
@@ -973,6 +985,15 @@ export const _Worker = Resource(
     // Use the provided name
     const workerName = props.name ?? id;
 
+    if (props.projectRoot) {
+      logger.warn("projectRoot is deprecated, use cwd instead");
+      props.cwd = props.projectRoot;
+    }
+
+    const cwd = props.cwd ? path.resolve(props.cwd) : process.cwd();
+    const relativeCwd =
+      cwd === process.cwd() ? undefined : path.relative(process.cwd(), cwd);
+
     const compatibilityDate =
       props.compatibilityDate ?? DEFAULT_COMPATIBILITY_DATE;
     const compatibilityFlags = props.compatibilityFlags ?? [];
@@ -985,7 +1006,7 @@ export const _Worker = Resource(
         upsertDevCommand({
           id,
           command: props.dev.command,
-          cwd: props.dev.cwd ?? process.cwd(),
+          cwd: props.dev.cwd ?? cwd,
           env: props.env ?? {},
         });
         return this({
@@ -993,6 +1014,7 @@ export const _Worker = Resource(
           id,
           entrypoint: props.entrypoint,
           name: workerName,
+          cwd: relativeCwd,
           compatibilityDate,
           compatibilityFlags,
           format: props.format || "esm", // Include format in the output
@@ -1036,6 +1058,7 @@ export const _Worker = Resource(
           {
             ...props,
             entrypoint: props.entrypoint,
+            cwd,
             compatibilityDate,
             compatibilityFlags,
           },
@@ -1134,6 +1157,7 @@ export const _Worker = Resource(
           (await bundleWorkerScript({
             name: workerName,
             ...props,
+            cwd,
             compatibilityDate,
             compatibilityFlags,
           }));
@@ -1152,6 +1176,7 @@ export const _Worker = Resource(
         id,
         entrypoint: props.entrypoint,
         name: workerName,
+        cwd: relativeCwd,
         compatibilityDate,
         compatibilityFlags,
         format: props.format || "esm", // Include format in the output
@@ -1182,6 +1207,7 @@ export const _Worker = Resource(
         (await bundleWorkerScript({
           ...props,
           name: workerName,
+          cwd,
           compatibilityDate,
           compatibilityFlags,
         }));
@@ -1358,30 +1384,32 @@ export const _Worker = Resource(
     };
 
     if (this.phase === "delete") {
-      // Delete any queue consumers attached to this worker first
-      await deleteQueueConsumers(api, workerName);
+      // only delete the worker if it's not a version
+      if (!props.version) {
+        // Delete any queue consumers attached to this worker first
+        await deleteQueueConsumers(api, workerName);
 
-      await withExponentialBackoff(
-        () =>
-          deleteWorker(api, {
-            workerName,
-            namespace:
-              typeof props.namespace === "string"
-                ? props.namespace
-                : props.namespace?.namespaceId,
-            url: this.output.url,
-          }),
-        (err) =>
-          (err.status === 400 &&
-            err.message.includes(
-              "is still referenced by service bindings in Workers",
-            )) ||
-          err.status === 500 ||
-          err.status === 503,
-        10,
-        100,
-      );
-
+        await withExponentialBackoff(
+          () =>
+            deleteWorker(api, {
+              workerName,
+              namespace:
+                typeof props.namespace === "string"
+                  ? props.namespace
+                  : props.namespace?.namespaceId,
+              url: this.output.url,
+            }),
+          (err) =>
+            (err.status === 400 &&
+              err.message.includes(
+                "is still referenced by service bindings in Workers",
+              )) ||
+            err.status === 500 ||
+            err.status === 503,
+          10,
+          100,
+        );
+      }
       return this.destroy();
     }
 
@@ -1435,7 +1463,7 @@ export const _Worker = Resource(
                 : customDomain.zoneId,
             adopt:
               typeof customDomain === "string"
-                ? false
+                ? (props.adopt ?? false)
                 : (customDomain.adopt ?? props.adopt),
           });
         }),
@@ -1464,7 +1492,7 @@ export const _Worker = Resource(
               typeof routeConfig === "string" ? undefined : routeConfig.zoneId, // Route resource will handle inference if not provided
             adopt:
               typeof routeConfig === "string"
-                ? false
+                ? (props.adopt ?? false)
                 : (routeConfig.adopt ?? props.adopt),
             accountId: props.accountId,
             apiKey: props.apiKey,
@@ -1507,6 +1535,7 @@ export const _Worker = Resource(
       id,
       entrypoint: props.entrypoint,
       name: workerName,
+      cwd: relativeCwd,
       compatibilityDate,
       compatibilityFlags,
       format: props.format || "esm", // Include format in the output